simplified checkbox, added round support
